+++ StoreKitUnitTests/StoreTransactionTests.swift
@@ -41,7 +41,7 @@ class StoreTransactionTests: StoreKitConfigTestCase {
     func testSK2DetailsWrapCorrectly() async throws {
         try AvailabilityChecks.iOS15APIAvailableOrSkipTest()
 
-        let sk2Transaction = try await self.simulateAnyPurchase()
+        let sk2Transaction = try await self.createTransactionWithPurchase()
 
         let transaction = StoreTransaction(sk2Transaction: sk2Transaction)
 
@@ -78,37 +78,6 @@ class StoreTransactionTests: StoreKitConfigTestCase {
 
         let transaction = StoreTransaction(sk1Transaction: sk1Transaction)
         expect(transaction.quantity) == 1
-    }
-
-}
-
-private extension StoreTransactionTests {
-
-    static let productID = "com.revenuecat.monthly_4.99.1_week_intro"
-
-    @MainActor
-    @available(iOS 15.0, tvOS 15.0, watchOS 8.0, macOS 12.0, *)
-    func simulateAnyPurchase() async throws -> SK2Transaction {
-        let product = try await fetchSk2Product()
-        _ = try await product.purchase()
-
-        let latestTransaction = await product.latestTransaction
-        let transaction = try XCTUnwrap(latestTransaction)
-
-        switch transaction {
-        case let .verified(transaction):
-            return transaction
-        default:
-            XCTFail("Invalid transaction: \(transaction)")
-            fatalError("Unreachable")
-        }
-    }
-
-    @MainActor
-    @available(iOS 15.0, tvOS 15.0, watchOS 8.0, macOS 12.0, *)
-    private func fetchSk2Product() async throws -> SK2Product {
-        let products: [Any] = try await StoreKit.Product.products(for: [Self.productID])
-        return try XCTUnwrap(products.first as? SK2Product)
     }
 
 }
